+++ Banners/__init__.py
@@ -1,1 +1,1 @@
-from Banners import Banners
+from Banners.Banners import Banners
