+++ Home/__init__.py
@@ -1,5 +1,5 @@
-from HomePage import HomePage
-from Buttons import Buttons
-from Directories import Directories
-from Documents import Documents
-from TabsAtHome import TabsAtHome
+from Home.HomePage import HomePage
+from Home.Buttons import Buttons
+from Home.Directories import Directories
+from Home.Documents import Documents
+from Home.TabsAtHome import TabsAtHome
